+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- `Context.create_initial_response` (and by extension `Context.respond` for the initial
+  response specifically) will no-longer try to pass the attachment, component or embed as
+  the actual message content when passed for the `content` argument for REST-based
+  interaction commands.
 
 ## [2.11.3] - 2023-02-01
 ### Added